switched default functionality to fullblock off - data from outside block with fullblock on may be incorrect
git-svn-id: file:///home/pkawak/git/svn/tmp/AMDAT_trunk/AMDAT/trunk@81 25419694-f2fb-4e8e-af61-79680089b9de

+++ mean_square_displacement.h
@@ -14,10 +14,10 @@ class Mean_Square_Displacement: public Analysis
 {
     int n_times;
     float * msd;
-    int * weighting;
+    float * weighting;
     float * timetable;
     void initialize(System*);                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  
-    int atomcount;
+    float atomcount;
     
     /*internal calculation variables*/
     int currenttime, nexttime, currenttimegap;
